chore(firebase): set up Firebase integration and update package structure
