+++ .gitignore
@@ -29,6 +29,6 @@ render.experimental.xml
 
 # Google Services (e.g. APIs or Firebase)
 google-services.json
-
+base64_google_services.txt
 # Android Profiling
 *.hprof
