+++ app/src/main/java/com/github/swent/swisstravel/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.github.swent.swisstravel
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -12,12 +12,16 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import com.android.sample.resources.C
-import com.android.sample.ui.theme.SampleAppTheme
+import com.github.swent.swisstravel.resources.C
+import com.github.swent.swisstravel.theme.SampleAppTheme
+import com.google.firebase.Firebase
+import com.google.firebase.auth.auth
+import com.google.firebase.firestore.firestore
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
+    useEmulator()
     setContent {
       SampleAppTheme {
         // A surface container using the 'background' color from the theme
@@ -27,6 +31,13 @@ class MainActivity : ComponentActivity() {
               Greeting("Android")
             }
       }
+    }
+  }
+
+  private fun useEmulator(isOnline: Boolean = false) {
+    if(isOnline) {
+      Firebase.firestore.useEmulator("10.0.2.2", 8080)
+      Firebase.auth.useEmulator("10.0.2.2", 9099)
     }
   }
 }
