+++ app/src/main/java/com/github/swent/swisstravel/SecondActivity.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.github.swent.swisstravel
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -12,8 +12,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import com.android.sample.resources.C
-import com.android.sample.ui.theme.SampleAppTheme
+import com.github.swent.swisstravel.resources.C
+import com.github.swent.swisstravel.theme.SampleAppTheme
 
 class SecondActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
