+++ app/src/main/java/com/github/swent/swisstravel/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.github.swent.swisstravel.theme
 
 import androidx.compose.ui.graphics.Color
