+++ app/src/main/java/com/github/swent/swisstravel/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.github.swent.swisstravel.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
