+++ build.gradle.kts
@@ -2,4 +2,6 @@
 plugins {
     alias(libs.plugins.androidApplication) apply false
     alias(libs.plugins.jetbrainsKotlinAndroid) apply false
+    //firebase google plugins
+    alias(libs.plugins.gms) apply false
 }
